playlist-service: move grpc handlers to handlers.go file
